+++ src/schema/v2/homeView/__tests__/HomeViewSection.test.ts
@@ -528,95 +528,6 @@ describe("HomeViewSection", () => {
     })
   })
 
-  describe("News", () => {
-    it("returns correct data", async () => {
-      const query = gql`
-        {
-          homeView {
-            section(id: "home-view-section-news") {
-              __typename
-              component {
-                title
-                type
-                behaviors {
-                  viewAll {
-                    buttonText
-                    href
-                    ownerType
-                  }
-                }
-              }
-
-              ... on HomeViewSectionArticles {
-                articlesConnection(first: 3) {
-                  edges {
-                    node {
-                      title
-                      href
-                    }
-                  }
-                }
-              }
-            }
-          }
-        }
-      `
-
-      const articles = [
-        {
-          title: "Bored apes stolen",
-          slug: "stolen-apes",
-        },
-        {
-          title: "More apes stolen",
-          slug: "more-apes",
-        },
-      ]
-
-      const context = {
-        articlesLoader: jest.fn().mockReturnValue({
-          count: articles.length,
-          results: articles,
-        }),
-      }
-
-      const { homeView } = await runQuery(query, context)
-
-      expect(homeView.section).toMatchInlineSnapshot(`
-        Object {
-          "__typename": "HomeViewSectionArticles",
-          "articlesConnection": Object {
-            "edges": Array [
-              Object {
-                "node": Object {
-                  "href": "/article/stolen-apes",
-                  "title": "Bored apes stolen",
-                },
-              },
-              Object {
-                "node": Object {
-                  "href": "/article/more-apes",
-                  "title": "More apes stolen",
-                },
-              },
-            ],
-          },
-          "component": Object {
-            "behaviors": Object {
-              "viewAll": Object {
-                "buttonText": "More in News",
-                "href": "/news",
-                "ownerType": "marketNews",
-              },
-            },
-            "title": "News",
-            "type": "ArticlesCard",
-          },
-        }
-      `)
-    })
-  })
-
   describe("HomeViewSectionTasks", () => {
     beforeAll(() => {
       mockIsFeatureFlagEnabled.mockImplementation((flag: string) => {
